feat(panels): migrate CCModal/SettingsModal to shared panels + component split plan
- Rewrite TargetRow/TargetsPanel to full-featured version (type selector,
  viewer icon picker + QR code, noBatch, format selector)
- Rewrite TranslationPanel to correct data shape {id, enabled, lang, target, format?}
  matching translationConfig.js; vendor picker, libre URL/key fields
- Rewrite RelaySlotRow/RelayPanel with canonical field names (targetType,
  youtubeKey) and 'Use original' checkboxes for scale/fps/bitrates
- CCModal: remove local TargetRow/TranslationRow; use TargetsPanel/TranslationPanel
- SettingsModal: add 'stream' tab to TABS; remove local RelayRow/DskRtmpUrlField;
  use RelayPanel with diff-based localStorage writes
- Fix wizard readLocalSettings/saveWizard/SetupWizardPage/ReviewSummary
  to use relayList with canonical field names
- Add docs/plans/plan_component_split.md: prioritized analysis of 9 monolithic
  components (CCModal, DskEditorPage, AudioPanel, BroadcastModal, SidebarLayout,
  DskViewportsPage, CaptionsModal, ControlsPanel, Production pages)

Agent-Logs-Url: https://github.com/jsilvanus/live-captions-yt/sessions/31f706f2-5de9-45b4-a486-44c5454053c8
